+++ src/guibbon/__init__.py
@@ -23,7 +23,7 @@ from .widgets.slider_widget import SliderWidget, CallbackSlider
 from .widgets.treeview_widget import TreeviewWidget, CallbackTreeview, TreeNode
 from .widgets.widget import WidgetInterface
 
-__version__ = "0.4.0-dev"
+__version__ = "0.4.0"
 
 
 def compute_version_info():